fix: Add more friendly tips
